added a loading spinner into pages of our site
