+++ src/Components/Team/TeamCards.jsx
@@ -1,36 +1,56 @@
-import React from "react";
+import React, { useState ,useEffect } from "react";
+import { Cards } from './Cards';
+import { Spinner } from "../spinner/spinner";
+import "./teamcards.css" ;
 
-const TeamCards = ({ teamList }) => {
+const TeamCards = () => {
 
+  const CoreTeamList = [
+    {
+      img: "https://avatars.githubusercontent.com/u/58720128?v=4",
+      name: "Shreeram Shanmugasundaram",
+      des: "Web Developer",
+      linkedin: "/",
+      github: "/",
+      color: "#ea4335",
+    },
+  
+  {
+    img:"https://avatars.githubusercontent.com/u/58720128?v=4",
+    name:"Aditya Gaikwad",
+    des :"Web Developer",
+    linkedin :"/",
+    github:"/",
+    color:"#34A853"
+  },{
+    img:"https://avatars.githubusercontent.com/u/58720128?v=4",
+    name:"Sneha Pujari",
+    des :"Web Developer",
+    linkedin :"/",
+    github:"/",
+    color:"#F9BC15"
+  },{
+    img:"https://avatars.githubusercontent.com/u/58720128?v=4",
+    name:"Kush Agrawal",
+    des :"Web Developer",
+    linkedin :"/",
+    github:"/",
+    color:"#4285F4"
+  
+  }]
+
+  const [loading , setLoading] = useState(true) ;
+  useEffect(() => {setTimeout(() => {setLoading(false);}, 2000);}, []);
   return (
-    <section className="center-div Team">
-      {teamList.map((item, index) => {
-        return (
-          <div className="Team_Card center-content" key={index}>
-            
-            <div className="SlideUp_Div" style={{background:item.color}}></div>
-            <img
-              src={item.img}
-              alt={item.name}
-              style={{ borderColor: item.color }}
-            />
-            <div className="Team_Card_Bottom">
-              <h3>{item.name}</h3>
-              <p>{item.des}</p>
-              <div className="Team_Social">
-                <a href={item.linkedin}>
-                  <i className="fa-brands fa-github"></i>
-                </a>
-                <a href={item.github}>
-                  <i className="fa-brands fa-linkedin"></i>
-                </a>
-              </div>
-            </div>
-
-          </div>
-        );
-      })}
-    </section>
+    <div className="teamcards">
+      {
+        loading ? (
+          <Spinner />
+        ) : (
+          <Cards teamList={CoreTeamList} />
+        )
+      }
+    </div>
   );
 };
