+++ src/Components/Team/teamcards.css
@@ -1,0 +1,110 @@
+.teamcards {
+    flex: auto;
+    flex-direction: column;
+    justify-items: center;
+    align-items: center ;
+    width: auto; 
+}
+.Team {
+    /* background-color: #fff; */
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
+    gap: 1rem;
+    margin: 2rem auto;
+  }
+  .Team_Card a {
+    text-decoration: none;
+  }
+  .Team_Card {
+      position: relative;
+      max-width: 200px;
+      margin: 0.5rem auto;
+    overflow: hidden;
+    border-radius: 10px;
+  
+  }
+  .SlideUp_Div{
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    background-color: #ea4335;
+    top: 100%;
+    transition: all ease-in-out 0.2s;
+    border-radius: 10px;
+  
+  
+  }
+  .Team_Card:hover .SlideUp_Div{
+    top: 0;
+  z-index: -1;
+  }
+  .Team_Card img {
+      display: block;
+    border-radius: 50%;
+    width: 100%;
+    border: solid #ea4335 5px;
+    object-position: center;
+    transition: all ease-in-out 0.2s;
+    
+  }
+  .Team_Card h3 {
+    margin-top: 1rem;
+    font-size: 20px;
+    color: #212529;
+    font-weight: 500;
+    letter-spacing: 0.5px;
+  }
+  .Team_Card p {
+    margin-top: 0.5rem;
+    font-size: 14px;
+    letter-spacing: 0.5px;
+  }
+  .Team_Social {
+    display: flex;
+    justify-content: center;
+  }
+  .Team_Card i {
+    display: inline-block;
+    margin: 1rem 0.4rem;
+    font-size: 20px;
+    width: 35px;
+    height: 35px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+  .Team_Card:hover {
+    border-radius: 10px;
+    transition: all ease-in-out 0.5s;
+    /* background-color: #ea4335; */
+  }
+  .Team_Card:hover img {
+    border-radius: 0%;
+    border: 0px solid #fff;
+  }
+  .Team_Card:hover h3 {
+    color: #fff;
+  }
+  .Team_Card:hover p {
+    color: #f2f2f2;
+  }
+  .Team_Card:hover i {
+    color: white;
+  
+  }
+  .Team_Card i:hover {
+    background-color: #212529;
+    color: white;
+    border-radius: 50%;
+  }
+  
+  .Team_Header{
+      padding: 2rem ;
+  }
+  
+  @media screen and (max-width: 800px) {
+    .Team{
+      grid-template-columns: repeat(2, 1fr);
+    }
+  
+  }
