+++ src/Components/spinner/spinner.css
@@ -1,0 +1,49 @@
+
+.load{
+    font-weight: 600;
+    background: linear-gradient(to right, from-red-600, to-yellow-300);
+    color: #50d71e;
+    padding-left: 0.5rem; 
+    padding-right: 0.5rem;
+    border-radius: 0.375rem;
+    font-size: 1.125rem;
+ }
+
+.spinner {
+    width: 56px;
+    height: 56px;
+    display: grid;
+    border: 4.5px solid #0000;
+    border-radius: 50%;
+    border-color: #fcff00 #0000;
+    animation: spinner-e04l1k 1s infinite linear;
+ }
+ .spinner::before, .spinner::after {
+    content: "";
+    grid-area: 1/1;
+    margin: 2.2px;
+    border: inherit;
+    border-radius: 50%;
+ }
+ .spinner::before {
+    border-color: #f70000 #0000;
+    animation: inherit;
+    animation-duration: 0.5s;
+    animation-direction: reverse;
+ }
+ .spinner::after {
+    margin: 8.9px;
+ }
+ @keyframes spinner-e04l1k {
+    100% {
+       transform: rotate(1turn);
+    }
+ }
+ .spin{
+    display: flex;
+    flex :auto ;
+    align-items: center;
+    flex-direction: column;
+    margin-top: 16px;
+    gap: 16px;
+ }
