+++ src/Components/spinner/spinner.jsx
@@ -1,0 +1,11 @@
+import React from 'react' ;
+import "./spinner.css" ;
+
+export const Spinner = () => {
+  return (
+    <div className="spin">
+            <div className="spinner"></div>
+            <p className="load">Loading...</p>
+    </div>
+  )
+}
